♻️ refactor(app): separate project view logic
- move project-related logic to a new ProjectView component
- simplify App component by delegating rendering to ProjectView

🐛 fix(app): hide waveform and controls without project

- ensure waveform and controls are not displayed when no project is loaded
- add conditional rendering to handle the absence of a current project

Fixes #25
